joid: Update verify jobs and introduce Multijob and verify phases
This is the first step towards improved Commit Gating and this
change does not impact the current way of working for Joid. This
means that no additional verification activities will be done for
changes/patches coming to Gerrit for Joid as of yet.

Because of this, the execute shells contain nothing but simple echos.
Functest job hasn't been created/connected either since we don't do
full commit verification.

This change
- switches verify job to Multijob project type since Freestyle does
  not support aborting all the downstream jobs when a new patch for
  same change comes to Gerrit for existing change. When Gerrit Trigger
  aborts the upstream verify job that was triggered for the previous
  patch of the same change, Multijob then aborts the downstream
  phase jobs, not leaving orphaned phase jobs behind.
- creates phase jobs to do limited verification activities in order
  to provide more granular feedback. Earlier phases fail the patch
  verification in order to provide feedback for the failed phase
  rather than whole basic, deploy, and smoke test.

Change-Id: If6723b06bb937329e104ca85865b0d750c45e91e
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>