fuel: Run ovs scenarios on ericsson-pod2
Fuel ovs scenarios require to have range of vlans trunked
on interface used for private vm networks. This configuration
is currently not available on LF POD2 so we have to have an exception
for these scenarios unless LF POD2 is configured accordingly.

This is kind of a messy workaround until this is handled properly.

Change-Id: I0e1a4cadaa318a1f353f82b5f6c8abb044d61af9
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>