Remove GIT_BASE parameter from slave parameters
All the slaves can clone repos from OPNFV Gerrit via https. As a
result of this, the previously introduced GIT_BASE parameter can
now be removed from slave-params.yml file to keep it clean and tidy
and prevent questions.

See the build history of the below job for clone testing on all the slaves
labelled with opnfv-sysinfo.
    https://build.opnfv.org/ci/job/repo_clone_test/

This change can only be submitted once the GIT_BASE env var on OPNFV Jenkins
Global Configuration is switched to https.

If any job fails on a slave due to issues with the cloning, it needs to
be fixed immediately.

Change-Id: I9620b9a6a9abed3696f1c917aa392a720f5b1cb0
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>