Update patch set 1
Patch Set 1: Code-Review+2

I think it would be better to override GS_URL slave-params.yml rather than hardcoding in to script since slave-params.yml is exactly for this type of stuff.

But it is good enough for the timebeing in order to prevent side effects that might occur due to change of the parameter ordering. One more thing to take a look again after the release.

Patch-set: 1
Label: Code-Review=+2