Update patch set 2
Patch Set 2: Code-Review-1

I don't think this covers enough possibilities.  What if the state of the host is already broken because the iptables rules are left on it.  We need a part of the clean at the beginning to remove iptables rules that are leftover.  For now, the mock detached stuff is breaking the daily.  I am going to revert releng patches that do mock detached stuff and we can try it again later.

Patch-set: 2
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1