Update patch set 1
Patch Set 1:

As I commented on the other patch (installer-params.yml I think), you will need to switch to {installer}-defaults some point in time. I submitted the other patch in order for you to move on with your work but I really prefer if you can switch to {installer}-params if it is possible.

Patch-set: 1