Update patch set 1
Patch Set 1:

Hi Fatih,

So the only real change here is the addition of the main.yml file. Your commit message does not explain why this change is necessary though. Could you add that so we remember in the future in case someone messes with this file? Moreover, is this something that could potentially be upstreamed? Finally, if the rest of the changes (file perms and removal of .gitkeep) are necessary, could you also document them in the commit message?

Patch-set: 1
Label: Code-Review=0