Update patch set 1
Patch Set 1:

> Moreover, is this something that could potentially be upstreamed? 

Markos,

This might not be upstreamed since bifrost really doesn't support VMs or they don't have use case for VMs.

The reason for us needing increased timeout is that when we switch to Ubuntu, this timeout value does not work anymore since it takes much more time to provision VMs with Ubuntu comparing cirros.

I'll update the commit message and include above info and the other changes in it.

Apart from this, we are just updating whatever is there to move forward and see how it works since we are still in prototyp/POC phase. I think we will need to create our own playbooks,etc and keep them in OPNFV instead of working with modified versions of the files.

Patch-set: 1
Label: Verified=0