Update patch set 1
Patch Set 1:

> As I commented on the other patch (installer-params.yml I think),
 > you will need to switch to {installer}-defaults some point in time.
 > I submitted the other patch in order for you to move on with your
 > work but I really prefer if you can switch to {installer}-params if
 > it is possible.
I think I'm starting to understand the hierarchy of the ymls now ...
How about we change the current to commit to use {installer}-defaults, and revert the other commit? We only want the INSTALLER_TYPE passed at ISO build time for now, and I guess "armand-defaults" is not really the way to go for that, but a simple {installer}-defaults include would have been enough.
What do you think?

Patch-set: 1