Update patch set 1
Patch Set 1:

That would be good. 

Instead of reverting the other one, just fix both in this change.

Just to let others know; the order is like this for daily jobs;

    -{installer}-defaults
    -{pod}-defaults

so this gives you possibility to override installer defaults based on the pod you are running stuff if any default parameter needs to be different for some reason.

The order is not that crucial for build jobs.

Patch-set: 1
Label: Code-Review=0