Update patch set 1
Patch Set 1: Code-Review+1

yes as joid installer creates ext-net network as part of post install script. But this parameter is referred by yardstick to reference in their heat templates. As we removed this in earlier patch so all yardstick cases are failing. So this patch is needed for yardstick not for joid as joid uses ext-net all the scenarios.

Patch-set: 1
Reviewer: Gerrit User 602 <602@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1