Update patch set 3
Patch Set 3: Code-Review+2

It is fine
we probably could create common libraries / image folder so for the evolution we will stay consistant?
Shall we do it now, or later?
I would suggest that now will be easier but the best way is probably to first merge then consider was it common and refactor functest and yardstick at the same time to use common libraries.

Patch-set: 3
Label: Code-Review=+2