Update patch set 11
Patch Set 11: Code-Review-1

For doctor-verify.sh we have a job already doing that. It is called 'functest-suite' where you can tell which test case you want to call. 
https://git.opnfv.org/cgit/releng/tree/jjb/functest/functest-ci-jobs.yml#n288

When calling that job with the parameter FUNCTEST_SUITE_NAME=doctor it will do the same as you wrote in doctor-verify.sh.

I think we should reuse it instead of writing code that does the same. That way we have only 1 place to maintain how we create the container and the env variables that have to be passed to it. I give -1 for this reason.

Is it possible to call that job from doctor.yaml?

Patch-set: 11
Label: Code-Review=-1