infra: Setup 3rd party CI job to verify bifrost patches in OPNFV
This change creates job to run patchset verification for openstack/bifrost.
The verification is done by checking out the patch from upstream bifrost
repo and combining opnfv bifrost stuff with it to see if any change that
is proposed upstream breaks opnfv bifrost.

The verification is done only for trusty and this can be expanded to support
centos7 and suse as well.

Due to change in slave label, impacted files are also updated.

Change-Id: Id34b7ddcf5e3f424a5999aa1a9fa69ff53d95fc4
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>