Update patch set 4
Patch Set 4:

We may be able to abandon some of these, if we know the input is trusted (for example you declare the input to eval as a variable inside your code e.g. cmd = 'python print ("hello)' 

The main danger with eval, is if it accepts input that is crafted externally (like via a webform).

I know have free time today, so I will help review these).

Patch-set: 4
Label: Verified=0