prototypes: bifrost: Check we are running the destroy-env.sh as root
Nearly all operations in the destroy-env.sh script need to be executed
as root so make sure that's the case before we try to do anything.
This also fixes the dnsmasq.leases path and we check that the file
exists before we try to wipe it. The dnsmasq is also restarted to take
the empty leases file into consideration. Finally, we improve the
documentation to run the necessary commands using sudo.

Change-Id: I56df821ca8a93307117a94a11abd0867b27d8443
Signed-off-by: Markos Chandras <mchandras@suse.de>