Update patch set 1
Patch Set 1:

(1 comment)

Mei, we are ready to add this into compass ci resource pool. Please fix the inline comment and make sure stuff can be run on this POD.

Once you amend the change and the release is out, the label will be added and this patch will be submitted.

Patch-set: 1