Remove GIT_BASE parameter from slave parameters
All the slaves can clone repos from OPNFV Gerrit via https.
As a result of this, the previously introduced GIT_BASE parameter
can now be removed from slave-params.yml file to keep it clean and
tidy and prevent questions.

This change can only be submitted once the GIT_BASE env var on OPNFV
Jenkins Global Configuration is switched to https. If any job fails
on a slave due to issues with the cloning, it needs to be fixed immediately.

Change-Id: Ie40c670a4f7db1d6794a5afb2d3ae4140c41fae8
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>