jjb: infra: Fix ownership for the cache directory
Fixes the following error when the ${HOME}/.cache permissions are bad:

The directory '/home/jenkins/.cache/pip/http' or its parent directory is
not owned by the current user and the cache has been disabled. Please
check the permissions and owner of that directory. If executing pip with
sudo, you may want sudo's -H flag.

Change-Id: I16310817e1f36bd8086e196ef61ec10c09ad7366
Signed-off-by: Markos Chandras <mchandras@suse.de>