fuel, armband: build, deploy: export TERM="vt220"
Since we do not actually parse our build/deploy logs for color chars,
set the terminal to "vt220".

This will silence quite a few of colorization special identifiers,
but NOT all of them (e.g. `tput sgr0` still prints a sequence).

For Fuel@OPNFV build process, we also need to workaround docker
terminal settings not being passed down when using a tty (see [1]),
which will be fixed in Docker 1.13+. This fix is handled in [2].

[1] https://github.com/docker/docker/issues/9299
[2] https://gerrit.opnfv.org/gerrit/#/c/22883/

Change-Id: I5f0c6eae33349ef81b67dd17e6df4e0af6274fab
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>