prototypes: bifrost: Ignore dnsmasq service failures
If this is the first time we run bifrost on a host or if we haven't
run bifrost since the last reboot, restarting the dnsmasq service will
fail because of conflicts with the dnsmasq process running by libvirt.
This is later on 'fixed' by bifrost by killing the one from libvirt.
Make sure that this spurious failure will not prevent the rest of the
cleanup process to move on.

Change-Id: Iebf37881570c27cb07cea3cecdb0fa8ac3fc511b
Signed-off-by: Markos Chandras <mchandras@suse.de>