Upload Generated Documentation to Correct Path
Documentation for Colorado 2.0 is not being uploaded to the correct
place. Instead of being uploaded to 'colorado/2.0/docs' documentation is
getting uploaded to 'colorado/2.0/'. Using `$gs_path` over `$GS_URL`
should fix this.

Further this should fix the failure from the gsutil command happening
when trying to set the correct Content-Type header on HTML files as it's
referencing a path that does not exist.

JIRA: DOCS-161

Change-Id: Ic732fda5cac54b1cc4834f2b44000ee87838f269
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>