Update patch set 1
Patch Set 1: Code-Review-1

OK if we use the CI production resources to validate new scenarios
but shall we not keep the scenario to continuously check there is no regression? is it not the definition of continutous integration...

Somehow I imagine moon should be still OK as it was OK in Colorado 1.0
the new developments affects the Master branch only
The colorado branch should not been in trouble
if trouble I would consider that as a regression (and then fixed) not the wish to skip the release

Moreover so far it only used virtual resources so I am not sure we will save lots of resources at the end

it could be interesting to define rules for that (not only for moon)
if a feature is candidate for release X, it shall be OK in CI during all the release cycle X.1, X.2, X.3 

see we have trouble with copper on last apex run (generic scenario) shall we exclude this scenario as well?
we had similar question on doctor/fuel on Colorado 1.0, it could be interesting to work on the way we manage that

Patch-set: 1
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1