vswitchperf: Utilize more CPU cores during build
Multiple CPU cores will be utilized during build of DPDK,
OVS and QEMU. It will speed up build significantly.
In case of VERIFY and MERGE jobs it can spare about 1/4
of time required for job execution.
Initial setup will use up to 10 cores for daily job,
which is executed on POD dedicated to vsperf project.
In case of VERIFY and MERGE jobs was number of cores
limited to 5 to evaluate impact on other jobs run
in parallel.

Change-Id: Ia0dc8f3d551c84ef2859f233c8442792e0b25c35
Signed-off-by: Martin Klozik <martinx.klozik@intel.com>
Reviewed-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>