utils: jenkins-jnlp-connect.sh: Drop -t from the monit service arguments
When executing the connection testing command as indicated in [1], -t is
added to the monit start-up script parameters. This prevents the jenkins
service from running in the background and create the pid file. Because
of that, monit is confused and reports the service as failed. We need to
filter out the '-t' argument in order to ensure that the service runs in the
background as expected.

[1]: http://artifacts.opnfv.org/octopus/brahmaputra/docs/octopus_docs/opnfv-jenkins-slave-connection.html#connecting-slaves-from-community-labs-to-opnfv-jenkins

Change-Id: I7371c4560311b84f56eff3461254264e1769ff7e
Signed-off-by: Markos Chandras <mchandras@suse.de>