utils: jenkins-jnlp-connect.sh: Make monit actions conditional
When executing the connection testing command as indicated in [1], -t is
added to the monit start-up script parameters. This prevents the jenkins
service from running in the background and create the pid file. Because
of that, monit is confused and reports the service as failed. We fix
this by moving all the monit related actions inside the 'monit'
conditional block.

[1]: http://artifacts.opnfv.org/octopus/brahmaputra/docs/octopus_docs/opnfv-jenkins-slave-connection.html#connecting-slaves-from-community-labs-to-opnfv-jenkins

Change-Id: I7371c4560311b84f56eff3461254264e1769ff7e
Signed-off-by: Markos Chandras <mchandras@suse.de>