Update patch set 2
Patch Set 2:

I have concerns about this - especially the case when you amend the patch. Here is the scenario.

you send a patch to gerrit and one of these 2 jobs get triggered. the other one gets put into the queue. while the first job is running, you amend the change and send a new patch. the running job gets aborted but I'm not sure if the one that that's waiting in the queue gets removed from the queue or not. this might result in 2 builds of the same job getting scheduled for the previous and current job.

My suggestion would be to stick to parent job and switch to the multijob project type so the job abort works correctly and we can avoid the deadlock situation for sure.

Patch-set: 2
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0