Update patch set 5
Patch Set 5:

The plan is not to have incremental TAGs for stable branches, which has been shown they are useles.. stable/colorado and stable/danube changes in the repo will generate the tag "stable" and that is the one to be used by the corresponding Jenkins Jobs. 

When we need to release OPNFV, I will have a different job that generates a given Tag (colorado.3.0, danube.1.0, ...) and it has to be run manually by a user.

Patch-set: 5