Update patch set 1
Patch Set 1:

btw the possible reason for the centos7 failures is that we use the 'centos' instead of 'centos-minimal' dib element which means that we use a much bigger rootfs for the centos VMs so it takes longer to install it on the disk and boot the VMs. However, even the minimal images can grow bigger overtime so it's best if we are a bit proactive here.

Patch-set: 1