3rd party ci: Create jobs for ODL netvirt testing in OPNFV
This change creates basic job structure for ODL netvirt testing in OPNFV.
The job should be triggered once the below job is completed
    https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-carbon
and fetch the artifact that is already built by the job. The build result codes
OPNFV is interested in are SUCCESSFUL and UNSTABLE.

Once the job gets triggered, the build URL will be fetched from ODL Gerrit using
Gerrit REST API and the artifact will be downloaded from ODL Jenkins.

The downloaded artifacts will then be installed on predeployed Apex environment
and tested.

The reason why we use Gerrit REST API is that the OPNFV Jenkins has an older version
of Gerrit Trigger Plugin which lacks feature that injects review comments as environment
variables. Upgrading Gerrit Trigger Plugin will not help either since OpenStack JJB
does not support the new version of Gerrit Trigger Plugin and does not generate the
new job XML.

Change-Id: I7ff8c5809dcb2e8ffa5c7c5aeb9cdf70e4e7e2f4
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>