3rd party ci: Create jobs for ODL netvirt testing in OPNFV
This change creates basic job structure for ODL netvirt testing in OPNFV.
The job should be triggered upon successful completion of
    https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-carbon
and fetch the artifact that is already built by the job.

Once the job gets triggered, the comment will be fetched from environment and
the artifacts built and archived on ODL Jenkins will be downloaded to the OPNFV
slave, installing it on predeployed Apex environment.

Change-Id: I7ff8c5809dcb2e8ffa5c7c5aeb9cdf70e4e7e2f4
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>