Add possibility to build a Release Docker image
New flag that if it's not empty, it will build the
image with the given tag instead of using master/stable.
These release builds will need to be triggered manually
by a Jenkins user with rights to build jobs.
The respective projects take the responsability to
build the release docker image whenever they
consider it is the right time.

Change-Id: I49b087e80eb134b8b3ddef8399f61eeac9b29619
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>