functest,doctor: CLEAN_DOCKER_IMAGES
The builder/script 'functest-cleanup' will delete docker images
for functest, so that we can make sure the latest functest will
be used during the tests. But, it also makes a pod downloads
image every time CI triggered functest-suite.

This patch adds option 'CLEAN_DOCKER_IMAGES' (default=true),
so that the job owner can choose whether he/she wants to drop
the docker images or not depends on frequency or charactaristics
of the job.

Change-Id: I0b7affccb404fb8c4f0939b00d65e363ec62f7c3
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>