Update patch set 4
Patch Set 4:

Thanks for clarifying Fatih.  We are almost done with the process to create the snapshots.  I was going to make a POC apex jjb and trigger an odl commit based on a comment just to make sure it all works on an apex pod.  Will then work with your and Nikolas to make it work with 3rd party jobs after that.

Patch-set: 4