Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > Yes, I think the container img does not require instant building,
 > for this file is not update frequently.
 > If required truely, maybe we can integrate with gerrit trigger.

Yes, it is not update frequently. We can tolerate it because even if we have new code depends on newly created container, we can merge the dockerfile change first and merge the new code change later as a workaround. On the other hand, even with gerrit change trigger, this job may still not be able to build container in time unless we do some blocking between this job and other jobs.

Patch-set: 2