3rd party ci: Create job to bring up apex vms
The job to bring up apex vms will be run on ericsson-virtual5.
The rest of the jobs will run on ericsson-virtual6, which is virtual
jumphost for the apex environment.

Apart from this, the job can now be (re)triggered by putting keyword
opnfv-test on ODL Gerrit.

There are still flaws and things to improve with the jobs such as
finding out if the downloaded artifact is real or just a file not found
response from ODL Jenkins and so on but these will be fixed later.

Change-Id: Ie38e513df08c6c76eacee53a112c09171b1f6c7d
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>