Skip cloning netvirt repo and fail the job if artifact doesn't exist
We do not need to clone netvirt repo as we are only interested
in the artifact.

Netvirt artifacts are removed from ODL Jenkins due to ODL Jenkins
artifact retention policy. This change ensures that we do not proceed
with the rest of the jobs if the artifact does not exist.

Change-Id: I9d8af28f5829fc9688c76f116d0bad737a9ecab6
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>