Update patch set 4
Patch Set 4:

Shall we not keep Colorado runs until Danube is released...

Most of the people are currently installing Colorado version so it would make sense to keep a minimum Colorado CI chain and at least the capability to re-run it. It would be better to add Danube and reduce Colorado resource allocation than a full replacement, no?

Patch-set: 4