Update patch set 2
Patch Set 2: Code-Review+2

sorry for the long delay...
looks great

do you know if the police , images js or css are licences/third party. If so it should be put in a third party forlder (as I did for the html5 framework used for Colorado). The tool checking the licence will send us warning if the code is third party code...
 
Anyway great job!
Up to us to develop the api to feed the DB with real data...

BTW do you know if it would be possible to use the same framework to create the page to select the custom list (i;e. a page showing the different test cases, I assume so...)

Patch-set: 2
Label: Code-Review=+2