Update patch set 4
Patch Set 4:

we can't keep all the jobs and pods for colorado as the danube development is going on. dovetail needs to bring up its needs to infra wg and to the installers. 

I proposed you a way; find a pod and we can set up the jobs to run scenarios deploying colorado and running dovetail against it. I really don't know how else I could explain this - similar to the discussions we had last year and we still don't understand what each other is talking about; we have a development to support and we have dovetail - they are 2 different things for me.

Patch-set: 4
Label: Verified=0