Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Hey,

sorry to be negative here. I think the apex snapshot and the netvirt-pipeline should not be coupled so closely. Better would be that we have a seperate job which create new snapshots and this job here will consume these snapshots whenever there is a new one.

@Tim: for you last comment. We have already 6 parallel netvirt pipelines running. ericsson-virtual5 to 11 are the virtual jumphost/jenkins slaves for that. ericsson-virtual5 is the slave which brings up the environments and 6-11 will do the netvirt installation and the functest testing.

Patch-set: 6
Label: Code-Review=-1