WIP: Create weekly jobs to deploy colorado and run dovetail from master
This will not work in its current state as the PODs ericsson-pod1 and
ericsson-pod2 weren't there when colorado.3.0 was released. These
PODs have become operational recently and required updates for these
PODs are not available in securedlab stable/colorado. There are at least
7 changes that need to be cherrypicked to stable/colorado but cherrypick
fails with conflict. Ericsson lab admins need to be involved in order to
make necessary changes correctly.

https://gerrit.opnfv.org/gerrit/#/c/26827/
https://gerrit.opnfv.org/gerrit/#/c/26867/
https://gerrit.opnfv.org/gerrit/#/c/26881/
https://gerrit.opnfv.org/gerrit/#/c/27057/
https://gerrit.opnfv.org/gerrit/#/c/26967/
https://gerrit.opnfv.org/gerrit/#/c/27089/
https://gerrit.opnfv.org/gerrit/#/c/27159/

Apart from POD specific updates, the jobs and scripts have updates which
might not work as expected but this is minor issue comparing the POD
updates.

Change-Id: I77e1ccb0bfe737a97bdcbe52c6918f59e3b37cbf
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>