Update patch set 1
Patch Set 1:

I suggest to change the approach here. Instead of running openstack-ansible from the host, i'd also create a jumphost with bifrost, as we were doing now with puppet-infracloud, and run openstack-ansible from there.
By that way, you avoid creating vlans and extra bridges on the deployment host, and the deployment is cleaner and less disruptive (as this host may need to run another jobs, not needing these packages and network configs)

Patch-set: 1