Update patch set 1
Patch Set 1:

In compass, it requires the substitution. I don't have fuel environment so I don't know how the original openrc looks like. We have to make sure whether the replacement is required by one installer or by all the installers. If it is only for one installer, we can think to move it to the specific block instead of being a shared logic.

Patch-set: 1