Update patch set 1
Patch Set 1:

I somehow expected that response. So, to a certain degree, this is redundant. I asked Feng Pan, who is doing the Gluon Apex integration, if he prefers a stable link or to parse the properties.latest file. Of course he prefers the stable link. A stable link would also allow us to directly link to the latest version for instance from the wiki. Moreover, it is just a 100Kb file and now we only have a (n+1) duplicate. 

However, I am open to suggestions.

Patch-set: 1