Update patch set 1
Patch Set 1:

> I somehow expected that response. So, to a certain degree, this is
 > redundant. I asked Feng Pan, who is doing the Gluon Apex
 > integration, if he prefers a stable link or to parse the
 > properties.latest file. Of course he prefers the stable link. A
 > stable link would also allow us to directly link to the latest
 > version for instance from the wiki. Moreover, it is just a 100Kb
 > file and now we only have a (n+1) duplicate.
 > 
 > However, I am open to suggestions.

The properties files are the ones what matter, keeping the metadata regarding the artifact and my concern is not about the filesize but having common practice regarding how we handle artifacts. If we go this direction, we will lose the only traceability we currently have.

It is perfectly understandable why direct link to a file simplifies things but what would happen if
- you decide to change the artifact filename. It impacts everyone who consumes your files. But properties files will always be there.
- if the consumers need to fetch netready repo for some reason and checkout the corresponding version. It won't be possible to do that as the properties file and the latest artifact might be out of sync or the artifact with latest name do not have corresponding properties file.

You can include link to latest.properties file in your wiki with a single sentence: Please fetch the artifact the link OPNFV_ARTIFACT_URL points.

I have no other suggestion but to use properties file.

Patch-set: 1