jjb: infra: bifrost-upload-logs.sh: Generate a landing page
Create a basic web page to contain all the collected build logs. The
reason for creating this page is that the OPNFV web server does not list
the subdirectory contents so we need another way to expose all the
uploaded logs.

Moreover, hide the gsutils stdout output since it just adds extra noise
to the already massive console log. Finally, drop the 'popd' command
argument since it causes the command to fail and mark the whole build
as failed.

Change-Id: Iad25e5cf62fda650bad3c9d83671cde84c7467d8
Signed-off-by: Markos Chandras <mchandras@suse.de>