jjb: infra: bifrost-upload-logs.sh: Consolidate upload step to main job
In case of job failures, the upload builder is not executed so the
published links are dead. We move the upload step to the EXIT handler
which is always executed despite the main job exit code.

Moreover, The gsutil tool can compress the files itself but in such
a way that the web browser can still display the contents without
the need for the user to download and uncompress the file himself.
This makes debugging more convenient. We also make the whole upload
process somewhat less noisy. Finally, we include the
$GERRIT_PATCHSET_NUMBER in the title and hyperlink since we provide
and retain build results for each patchset instead of a single set
just for the latest one.

Change-Id: I43efd519a5743661d1f258aa338fb0769f59b6ed
Suggested-by: Julien Zhang <zhang.jun3g@zte.com.cn> # compression only
Signed-off-by: Markos Chandras <mchandras@suse.de>