jjb: infra: bifrost-upload-logs.sh: Use native gsutil compression
The gsutil tool can compress the files itself but in such a way that
the web browser can still display the contents without the need for
the user to download and uncompress the file himself. This makes
debugging more convenient. Moreover, we make the whole upload process
somewhat less noisy.

Change-Id: I43efd519a5743661d1f258aa338fb0769f59b6ed
Suggested-by: Julien Zhang <zhang.jun3g@zte.com.cn>
Signed-off-by: Markos Chandras <mchandras@suse.de>