jjb: infra: bifrost-verify.sh: Rework code for uploading the build log
gsutil does not support compression on streaming uploads so rework the
code to save the log into a temporary file first. Fixes the following
error:

CommandException: gzip compression is not currently supported on
streaming uploads. Remove the compression flag or save the streamed
output temporarily to a file before uploading.

Change-Id: I3976bf3271be3aff8bc7b656ac49de6986ca6c36
Signed-off-by: Markos Chandras <mchandras@suse.de>