jjb: infra: bifrost-verify.sh: Always create the landing page
Previously, the function would have returned if no bifrost logs were
present. However, this prevented the landing page from being generated
and uploaded since that happens at the very end of that function.
We rewored the code to always create a landing page even if there
are no bifrost logs to upload.

Change-Id: Ice557d0a42324a135d7f0e5ec94ec22811cc0ae7
Signed-off-by: Markos Chandras <mchandras@suse.de>