Add new jobs for doc merge and verify
New jobs for merge and verify added in tandem to ODLDocs.  Intent of
keeping these doc jobs  between ODL & OPNFV as similar as possible is
to keep the scope for common jobs and cross collaboration possible.
The archiving feature that is included in post build configuration of
ODLDoc jobs has been left out for now until CI Team can take it up.

Change-Id: Ifb494f1491f07ad0af39cd2644f895372b98edf3
Signed-off-by: ShubhamRathi <shubhamiiitbackup@gmail.com>