Add choosing strategy for git-scm-gerrit
Add choice for git-scm-gerrit between gerrit and default

/jjb/opnfvdocs/opnfvdocs.yml now has choosing-strategy: 'default'

We need default for docs merge so that we are not merging
the gerrit patchset, but instead the code as it is in
master after the merge

All other instances of git-scm-gerri  have been set to gerrit as
this was the default before. We may want to change this.
but I leave it as an excercise to the projects

Change-Id: I32c6b5f574f957088364177bb8387be71b1e3ec7
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>