jjb: infra: bifrost-verify.sh: Always print the GS URL
Previously, the GS URL was only printed when the job succeeded and we
were able to collect all bifrost logs. However, it's also helpful to be
able to see the URL when a job fails so move it outside of the 'if'
block. Moreover, improve the wording in other informative messages.

Change-Id: I080d5c1ff49de5abe50cc1049bb38f6e42123587
Signed-off-by: Markos Chandras <mchandras@suse.de>