Apex: Adds gate scenario job
Apex needs a way to run a gate job after basic scenarios (nosdn,
odl-nofeature) pass.  We need the capability to specify which scenario
should be the gate.  This patch allows us via a comment to specify
'start-gate-scenario: <scenario>' which will start a gate job to
validate that scenario deploys and passes basic functest health check.

Some drawbacks of this approach include only being able to validate a
single scenario per comment + gerrit votes overriding (not true
workflow/pipeline plugin).  The positive is this is a fairly small CI
change to enable some much needed verification.  It will be the Apex
committers responsibility to issue this comment post Jenkins +1 approval
of pre-gate.

Change-Id: I9a1364bb1cece8b35488d496789a5e83c707b32d
Signed-off-by: Tim Rozet <trozet@redhat.com>