Update patch set 6
Patch Set 6:

(1 comment)

I am planning to create a class for RoleType or something similar in a new commit, this way we don't hardcode the roles. But need to figure out first what are all the possible roles from all the installers and how they are printed out with the corresponding commands.

Patch-set: 6