Apex,3rd_party: Fix workspace wiping
Workspace wiping was done in builders, which can be removed in favor of
wiping workspace via the 'scm' entry in the job.  For now we point at
apex since we require apex repo to clean the host properly.  We also
need sdnvpn for the python code to upgrade ODL.  Since we can only use a
single scm at a time, we use apex in the jjb, and then manually clone
sdnvpn in the builders that require it.

Change-Id: I52274fc8fdd6b6c052e396c2a63dab5abcbf764e
Signed-off-by: Tim Rozet <trozet@redhat.com>