CSIT: fixes for functest
Since we have no 'installer' for CSIT snapshot deployments, we need to
indicate to functest where the RC credentials are.  This patch adds
that, as well as modifying the IP configured for the host br-external
bridge to be the GW IP for the external neutron network.  This will
allow floating IPs to work with the Functest container.

Change-Id: Ie594de8cb2d7837abfad612b9ba49b0c02e7b2be
Signed-off-by: Tim Rozet <trozet@redhat.com>