Apex: Removes nosdn from verify
Since we now have gate jobs in order to verify functionality specific to
a scenario, we can eliminate running 2 pre-gate verify deploys with
nosdn and odl_l3.  The consensus from Apex team is that we gain little
validation with nosdn in comparison to odl_l3, so we can remove nosdn
and then trigger it as a specific gate job if we need to validate its
functionality.

Change-Id: I23367ad746cedd8b091aa4ffaf53249fef3c0eee
Signed-off-by: Tim Rozet <trozet@redhat.com>