Update patch set 1
Patch Set 1:

> same? https://gerrit.opnfv.org/gerrit/#/c/29231/

Not the same with 29231, the patch 29231 aims to verify the jjb yaml files and drives the unittest in module directory.

Here I use it to drive pep8 check and unittest of TestAPI.

The reason I want to introduce tox is because, it will really facilitate our work. There are a few checks needed when submitting a patch: pep8 check, unittest, and api-docs generation(Rohit implement the auto generation after merge, but it may fail to generate if we don't validate it before merge), for now it is executed manually one after another. Of course I can write a script to do it. but I don

Patch-set: 1