Update patch set 2
Patch Set 2:

> Hi please document the intended usage in opnfv-jjb-usage.rst
 > please remove the README.md, if you would like to add this place it
 > as an rst in docs/
 > 
 > I would prefer if setup.cfg etc were in a subdirectory and not in
 > the root of releng.

I will add the usage in opnfv-jjb-usage.rst;

OK,  will delete README.md, which is only for tox;

I think keep the setup.cfg in the top layer is better, which is defined for releng project and can be used to integrate any sub module in the releng. As you can see, the setup.cfg is just a framework, no actually requriements in it.

Patch-set: 2
Label: Verified=0