jjb: infra: bifrost: Enable disable-strict-forbidden-file-verification
Previously, if an upstream patchset contained a change for a file
listed in 'forbidden-file-paths' the job would not be triggered. This
is not desirable since such a patchset may contain important changes so
we enable the 'disable-strict-forbidden-file-verification' option which
triggers the job unless the patchset only contains changes for the
files listed in 'forbidden-file-paths'.

Change-Id: Ifa86d3a39b36375b2fd52b449e29c8dc757499b4
Signed-off-by: Markos Chandras <mchandras@suse.de>