Update patch set 1
Patch Set 1:

> > Patch Set 1: Code-Review+1
 > 
 > there is a link(https://gerrit.opnfv.org/gerrit/#/c/30219) in
 > commit-msg, and you can see why.

This commit is independent of the linked patchset so it may still be helpful to explain the reasoning. If you don't want to repeat the commit message you can wait for that commit to be accepted and then use the git hash as a reference so one does not need internet access to understand it in the future. It's just good git practice to document every commit as best as we can because if you look at this commit in two years time, you may not be able to understand it without going back to gerrit.

So neither -1 nor +1 from me here :)

Patch-set: 1