Adds cronjob for filling vnf table row entries
It uses github api calls to fill no_of_developers, no_of_stars and
versions entry. It is a bit hacky as it sets some parameters and parses
the header link line to get the number.
TODO lines_of_code entry as it is not provided directly by github api.

Change-Id: I6cba54ad48fbb6aa5f4823100a426abbb82fc447
Signed-off-by: Kumar Rishabh <shailrishabh@gmail.com>