jjb: infra: bifrost-cleanup-job: Use strings instead of dicts for triggers
JJB does not seem to handle new style all 'trigger-on' triggers when they
are in dictionary format. A patch is under consideration in
https://review.openstack.org/#/c/446225/ and until this one or another
solution is merged we convert these triggers to strings which work as
expected. With the previous code the resulted xml output was:

</triggerOnEvents>

and with the new one is

<triggerOnEvents>
 <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginChangeMergedEvent/>
 <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginChangeAbandonedEvent/>
</triggerOnEvents>

Link: https://review.openstack.org/#/c/446225/
Change-Id: I7fd1a6deca92e38ad222dc21e27aaa65a367368c
Signed-off-by: Markos Chandras <mchandras@suse.de>